+++ src/pages/shop/ShopPage.tsx
@@ -6,7 +6,7 @@ import {
   type NftGoodsItem,
 } from "../../apis/blockchain/blockchain";
 
-export default function NftCollectionPage() {
+export default function ShopPage() {
   const navigate = useNavigate();
   const [goods, setGoods] = useState<NftGoodsItem[]>([]);
   const [loading, setLoading] = useState(true);
@@ -32,13 +32,11 @@ export default function NftCollectionPage() {
       .finally(() => setLoading(false));
   }, [accessToken]);
 
-  // 판매된 항목은 "소유 중"으로 표시 (실제로는 owner 필드 기준 처리)
   const owned = goods.filter((g) => g.isSold);
   const available = goods.filter((g) => !g.isSold);
 
   return (
     <div className="flex h-full flex-col bg-white">
-      {/* 헤더 */}
       <div className="flex items-start justify-between px-4 py-3 border-b border-slate-100">
         <div className="flex items-center gap-2">
           <button className="flex flex-col gap-1 p-1" onClick={() => navigate("/campus")}>
@@ -56,7 +54,6 @@ export default function NftCollectionPage() {
         </div>
       </div>
 
-      {/* 본문 */}
       <div className="flex-1 overflow-y-auto px-4 py-4 space-y-6">
         {loading ? (
           <div className="flex items-center justify-center py-20">
@@ -64,7 +61,6 @@ export default function NftCollectionPage() {
           </div>
         ) : (
           <>
-            {/* MY BUILDINGS */}
             <section>
               <p className="text-xs font-semibold tracking-widest text-slate-400 mb-3">
                 MY BUILDINGS
@@ -105,7 +101,6 @@ export default function NftCollectionPage() {
               )}
             </section>
 
-            {/* AVAILABLE */}
             <section>
               <p className="text-xs font-semibold tracking-widest text-slate-400 mb-3">
                 AVAILABLE
@@ -135,7 +130,6 @@ export default function NftCollectionPage() {
                   </button>
                 ))}
 
-                {/* Coming soon 빈 슬롯 */}
                 {available.length % 2 !== 0 && (
                   <div className="rounded-xl border border-dashed border-slate-200 flex flex-col items-center justify-center aspect-square">
                     <p className="text-xs text-slate-300">Coming soon</p>
